% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.r
\name{clean_names}
\alias{clean_names}
\title{Clean names of an object.}
\usage{
clean_names(data)
}
\arguments{
\item{data}{Object, usually \code{data.frame}, with column names to clean.}
}
\value{
The input object \code{data} with clean column names.
}
\description{
The function imports the \code{clean_names} function from the \code{janitor} package.
}
\section{Warning}{

The function requires that the \code{janitor} is installed!
}

\examples{
data <- create_data()
names(data) <- gsub("_", ".", names(data))
clean_names(data)

}
\seealso{
\code{\link[janitor]{clean_names}}
}
