% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Count Occurrences of Unique Vector Elements}
\usage{
count(vec = NULL)
}
\arguments{
\item{vec}{(vector) vector containing elements to count}
}
\value{
(dataframe) two-column dataframe with as many rows as there are unique elements in the provided vector. First column is named "value" and includes the unique elements of the vector, second column is named "count" and includes the number of occurrences of each vector element.
}
\description{
Counts the number of occurrences of each element in the provided vector. Counting of NAs in addition to non-NA values is supported.
}
\examples{
# Count instances of vector elements
count(vec = c(1, 1, NA, "a", 1, "a", NA, "x"))

}
