\name{create_data}
\alias{create_data}
\title{Create data}
\usage{
create_data(n = 1000)
}
\arguments{
\item{n}{Number of observations to be simulated}
}
\value{
\code{data.frame} with columns profit, cogs, fixed_cost, rnd, assets, competition, quality, board_size, ceo_age, firm_value, female_ceo, industry, and country.
}
\description{
Generate simulated dataset
}
