% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_tri.R
\name{crop_tri}
\alias{crop_tri}
\title{Crop a Triangle from Data Object}
\usage{
crop_tri(data = NULL, drop_tri = "upper", drop_diag = FALSE)
}
\arguments{
\item{data}{(dataframe, dataframe-like, or matrix) symmetric data object to remove one of the triangles from}

\item{drop_tri}{(character) which triangle to replace with NAs, either "upper" or "lower"}

\item{drop_diag}{(logical) whether to drop the diagonal of the data object (defaults to FALSE)}
}
\value{
(dataframe or dataframe-like) data object with desired triangle removed and either with or without the diagonal
}
\description{
Accepts a symmetric data object and replaces the chosen triangle with NAs. Also allows user to choose whether to keep or drop the diagonal of the data object
}
