% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_chk.R
\name{date_chk}
\alias{date_chk}
\title{Check a Column for Non-Dates.}
\usage{
date_chk(data, col)
}
\arguments{
\item{data}{A data frame}

\item{col}{A column name from the provided data frame as a character vector}
}
\value{
A character vector
}
\description{
Any elements in the column that would be changed to NA if `as.Date()` is used on the column are returned. This is useful for quickly identifying only the "problem" entries of an ostensibly date column that is read in as a character.
}
\examples{
# Let's create some data that will be useful in demoing this function
loc <- c("LTR", "GIL", "PYN", "RIN")
time <- c('2021-01-01', '2021-01-0w', '1990', '2020-10-xx')
sites <- data.frame('site' = loc, 'visit' = time)

# Now we can use our function to identify bad dates
date_chk(data = sites, col = 'visit')
}
