% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_format_guess.R
\name{date_format_guess}
\alias{date_format_guess}
\title{Identify Probable Format for Ambiguous Date Formats.}
\usage{
date_format_guess(
  data = NULL,
  date_col = NULL,
  groups = TRUE,
  group_col = NULL,
  return = "dataframe",
  quiet = FALSE
)
}
\arguments{
\item{data}{(dataframe) object containing at least one column of ambiguous dates}

\item{date_col}{(character) name of column containing ambiguous dates}

\item{groups}{(logical) whether groups exist in the dataframe / should be used (defaults to TRUE)}

\item{group_col}{(character) name of column containing grouping variable}

\item{return}{(character) either "dataframe" or "vector" depending on whether the user wants the date format "guesses" returned as a new column on the dataframe or a vector}

\item{quiet}{(logical) whether certain optional messages should be displayed (defaults to FALSE)}
}
\value{
(dataframe or character) object containing date format guesses
}
\description{
In a column containing multiple date formats (e.g., MM/DD/YYYY, "YYYY/MM/DD, etc.) identifies probable format of each date. Provision of a grouping column improves inference. Any "unguessable" formats are flagged as "FORMAT UNCERTAIN" for human double-checking. This is useful for quickly sorting the bulk of ambiguous dates into clear categories for later conditional wrangling.
}
