% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_chk.R
\name{diff_chk}
\alias{diff_chk}
\title{Compare Difference Between two Vectors}
\usage{
diff_chk(old = NULL, new = NULL)
}
\arguments{
\item{old}{(vector) starting / original object}

\item{new}{(vector) ending / modified object}
}
\description{
Reflexively compares two vectors and identifies (1) elements that are found in the first but not the second (i.e., "lost" components) and (2) elements that are found in the second but not the first (i.e., "gained" components). This is particularly helpful when manipulating a dataframe and comparing what columns are lost or gained between wrangling steps. Alternately it can compare the contents of two columns to see how two dataframes differ.
}
