% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr_get_summary.r
\name{.hlpr_get_summary}
\alias{.hlpr_get_summary}
\alias{.hlpr_get_summary.numeric}
\alias{.hlpr_get_summary.logical}
\alias{.hlpr_get_summary.character}
\alias{.hlpr_get_summary.factor}
\title{get_summary}
\usage{
.hlpr_get_summary(input, ...)
}
\description{
get_summary
}
\keyword{internal}
