% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr_modular_models.r
\name{.hlpr_modular_models}
\alias{.hlpr_modular_models}
\alias{.hlpr_modular_models.param_glm}
\alias{.hlpr_modular_models.param_glmer}
\title{modular_models}
\usage{
.hlpr_modular_models(parameters, ...)
}
\description{
modular_models
}
\keyword{internal}
