% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr_modular_report.r
\name{.hlpr_modular_report}
\alias{.hlpr_modular_report}
\title{modular_report}
\usage{
.hlpr_modular_report(model_list, cv, show_cv, robust_type, var_cluster)
}
\description{
modular_report
}
\keyword{internal}
