% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms_ord.R
\name{nms_ord}
\alias{nms_ord}
\title{Publication-Quality Non-metric Multi-dimensional Scaling (NMS) Ordinations}
\usage{
nms_ord(
  mod,
  groupcol,
  title = NA,
  colors = c("#41b6c4", "#c51b7d", "#7fbc41", "#d73027", "#4575b4", "#e08214",
    "#8073ac", "#f1b6da", "#b8e186", "#8c96c6"),
  lines = rep(1, 10),
  leg_pos = "bottomleft",
  leg_cont = unique(groupcol)
)
}
\arguments{
\item{mod}{Object returned by `vegan::metaMDS()`}

\item{groupcol}{Column in the data that includes the groups}

\item{title}{Character string to use as title for plot}

\item{colors}{Vector of colors (as hexadecimal codes) of length >= group levels (currently *not* colorblind safe because of need for 10 built-in unique colors)}

\item{lines}{Vector of line types (as integers) of length >= group levels}

\item{leg_pos}{Legend position, either numeric vector of x/y coordinates or shorthand accepted by `graphics::legend()`}

\item{leg_cont}{Concatenated vector of desired legend entries. Defaults to unique entries in "groupcol" argument (this argument provided in case syntax of legend contents should differ from data contents)}
}
\description{
Produces Non-Metric Multi-dimensional Scaling (NMS) ordinations for up to 10 groups. Assigns a unique color for each group and draws an ellipse around the standard deviation of the points. Automatically adds stress (see `vegan::metaMDS()` for explanation of "stress") as legend title. Because there are only five hollow shapes (see `?graphics::pch()`) all shapes are re-used a maximum of 2 times when more than 5 groups are supplied.
}
\examples{
# Use data from the vegan package
utils::data("varespec", package = 'vegan')
resp <- varespec

# Make some columns of known number of groups
factor_2lvl <- c(rep.int("Trt1", (nrow(resp)/2)),
                 rep.int("Trt2", (nrow(resp)/2)))
factor_4lvl <- c(rep.int("Trt1", (nrow(resp)/4)),
                 rep.int("Trt2", (nrow(resp)/4)),
                 rep.int("Trt3", (nrow(resp)/4)),
                 rep.int("Trt4", (nrow(resp)/4)))
factor_6lvl = c(rep.int("Trt1", (nrow(resp)/6)),
                rep.int("Trt2", (nrow(resp)/6)),
                rep.int("Trt3", (nrow(resp)/6)),
                rep.int("Trt4", (nrow(resp)/6)),
                rep.int("Trt5", (nrow(resp)/6)),
                rep.int("Trt6", (nrow(resp)/6)))
factor_over <- (1:nrow(resp))

# And combine them into a single data object
data <- cbind(factor_over, factor_2lvl, factor_4lvl, factor_6lvl, resp)

# Actually perform multidimensional scaling
mds <- vegan::metaMDS(data[-c(1:4)], autotransform = FALSE, expand = FALSE, k = 2, try = 100)

# With the scaled object and original dataframe we can use this function
helpR::nms_ord(mod = mds, groupcol = data$factor_4lvl,
                title = '4-Level NMS', leg_pos = 'topright',
                leg_cont = c('1', '2', '3', '4'))

# And too many groups results in an informative error
helpR::nms_ord(mod = mds, groupcol = data$factor_over)

}
