% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_chk.R
\name{num_chk}
\alias{num_chk}
\title{Check a Column for Non-Numbers.}
\usage{
num_chk(data, col)
}
\arguments{
\item{data}{A data frame}

\item{col}{A column name from the provided data frame as a character vector}
}
\value{
A character vector
}
\description{
Any elements in the column that would be changed to NA if `as.numeric()` is used on the column are returned. This is useful for quickly identifying only the "problem" entries of an ostensibly numeric column that is read in as a character.
}
\examples{
# Create dataframe with a numeric column where some entries would be coerced into NA
spp <- c('salmon', 'bass', 'halibut', 'eel')
ct <- c(1, '14x', '_23', 12)
fish <- data.frame('species' = spp, 'count' = ct)

# Use `num_chk()` to return only the entries that would be lost
helpR::num_chk(data = fish, col = "count")
}
