% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderstatistics.R
\name{ParetoOrder}
\alias{ParetoOrder}
\title{Moments of order statistics for two-parameter Pareto}
\usage{
ParetoOrder(q = 1, r = 1, scale = 1, shape = 2, pow = 1)
}
\arguments{
\item{q}{The order of statistics (q-th out of r)}

\item{r}{The number of replicates to extract the q-th statistics from (q-th out of r)}

\item{scale}{The scale of two-parameter Pareto distribution}

\item{shape}{The shape of two-parameter Pareto distribution}

\item{pow}{The order of the moment to obtain statistics for (default pow=1 stands for expectation)}
}
\value{
The pow-th moment of q-th order statistics of r iid. copies of the
   r.v. distribued as the two-parameter Pareto with distribution function
   \deqn{F(x)=1-\left(\frac{scale}{x}\right)^{shape}, \;\; x\geq scale.}
}
\description{
ParetoOrder is the moments of order statistics from two-parameter Pareto distribution.
}
\examples{
ParetoOrder() # should be 2
}
\references{
J. S. Huang, “A Note on Order Statistics from Pareto Distribution,” Scandinavian Actuarial Journal, vol. 1975, no. 3, pp. 187–190, Jul. 1975, doi: 10.1080/03461238.1975.10405095.
}
\seealso{
[supportR::WeibullOrder()] for moments from three-parameter Weibull distribution.
}
