% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_ord.R
\name{pcoa_ord}
\alias{pcoa_ord}
\title{Publication-Quality Principal Coordinates Analysis (PCoA) Ordinations}
\usage{
pcoa_ord(
  mod,
  groupcol,
  title = NA,
  colors = c("#c51b7d", "#7fbc41", "#d73027", "#4575b4", "#e08214", "#8073ac",
    "#f1b6da", "#b8e186", "#8c96c6", "#41b6c4"),
  lines = rep(1, 10),
  leg_pos = "bottomleft",
  leg_cont = unique(groupcol)
)
}
\arguments{
\item{mod}{Object returned by `ape::pcoa()`}

\item{groupcol}{Column in the data that includes the groups}

\item{title}{Character string to use as title for plot}

\item{colors}{Vector of colors (as hexadecimal codes) of length >= group levels (currently *not* colorblind safe because of need for 10 built-in unique colors)}

\item{lines}{Vector of line types (as integers) of length >= group levels}

\item{leg_pos}{Legend position, either numeric vector of x/y coordinates or shorthand accepted by `graphics::legend()`}

\item{leg_cont}{Concatenated vector of desired legend entries. Defaults to unique entries in "groupcol" argument (this argument provided in case syntax of legend contents should differ from data contents)}
}
\description{
Produces Principal Coordinates Analysis (PCoA) ordinations for up to 10 groups. Assigns a unique color for each group and draws an ellipse around the standard deviation of the points. Automatically adds percent of variation explained by first two principal component axes parenthetically to axis labels. Because there are only five hollow shapes (see `?graphics::pch()`) all shapes are re-used a maximum of 2 times when more than 5 groups are supplied.
}
\examples{


# Use data from the vegan package
utils::data("varespec", package = 'vegan')
resp <- varespec

# Make some columns of known number of groups
factor_2lvl <- c(rep.int("Trt1", (nrow(resp)/2)),
                 rep.int("Trt2", (nrow(resp)/2)))
factor_4lvl <- c(rep.int("Trt1", (nrow(resp)/4)),
                 rep.int("Trt2", (nrow(resp)/4)),
                 rep.int("Trt3", (nrow(resp)/4)),
                 rep.int("Trt4", (nrow(resp)/4)))
factor_6lvl = c(rep.int("Trt1", (nrow(resp)/6)),
                rep.int("Trt2", (nrow(resp)/6)),
                rep.int("Trt3", (nrow(resp)/6)),
                rep.int("Trt4", (nrow(resp)/6)),
                rep.int("Trt5", (nrow(resp)/6)),
                rep.int("Trt6", (nrow(resp)/6)))
factor_over <- (1:nrow(resp))

# And combine them into a single data object
data <- cbind(factor_over, factor_2lvl, factor_4lvl, factor_6lvl, resp)

# Get a distance matrix from the data
dist <- vegan::vegdist(resp, method = 'kulczynski')

# Perform a PCoA on the distance matrix to get points for an ordination
pnts <- ape::pcoa(dist)

# Test the function for 4 groups
helpeR::pcoa_ord(pnts, data$factor_4lvl)

# Look what happens if you go over the supported number of groups:
helpeR::pcoa_ord(pnts, data$factor_over)

}
