% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderstatistics.R
\name{PHOrder}
\alias{PHOrder}
\title{Moments of order statistics for Phase-type distribution PH(beta,B)}
\usage{
PHOrder(q = 1, r = 1, B = -1, beta = 1, pow = 1)
}
\arguments{
\item{q}{The order of statistics (q-th out of r)}

\item{r}{The number of replicates to extract the q-th statistics from (q-th out of r)}

\item{B}{The matrix describing Phase-type distribution}

\item{beta}{The initial probability vector describing Phase-type distribution}

\item{pow}{The order of the moment to obtain statistics for (default pow=1) - CURRENTLY UNUSED}
}
\value{
The pow-th moment of q-th order statistics of r iid. copies of the
   r.v. distribued as the Phase-type distribution PH\eqn{(\beta,B)} with density
   \deqn{F(x)=1-\beta e^{B x}\mathbf{e},}
   where \eqn{\mathbf{e}=(1,\dots,1)} is vector of ones.
}
\description{
PHOrder calculates the moments of order statistics from Phase-type distribution PH(beta,B).
}
\examples{
PHOrder() # should be 1 as the expectation of standard Exponential
PHOrder(1,10,-1,1) # should be 0.1 as the minimum of 10 standard Exponentially distributed r.v.'s
}
\references{
M. Bladt and B. F. Nielsen, Matrix-Exponential Distributions in Applied Probability, vol. 81. in Probability Theory and Stochastic Modelling, vol. 81. Boston, MA: Springer US, 2017. doi: 10.1007/978-1-4939-7049-0.
}
\seealso{
[supportR::WeibullOrder()] for moments from three-parameter Weibull distribution.
}
