% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edi.R
\name{read_edi}
\alias{read_edi}
\title{Read Entire Data Package from Environmental Data Initiative (EDI) by PASTA Identifier}
\usage{
read_edi(pasta_id = NULL, data_type = "csv")
}
\arguments{
\item{pasta_id}{(character) PASTA identifier. Can be found in the EDI page for a given data product just above the "Code Generation" buttons for various coding languages.}

\item{data_type}{(character) file extension of the data tables (i.e., non-metadata parts of data package). Currently only supports CSV format.}
}
\value{
(list) containing one element per table in data product and the final element is the XML metadata. List names are the PASTA identifiers for each specific object (note that these differ slightly from the PASTA identifier this function requires)
}
\description{
Reads in the data files and metadata (as an XML) stored on the Environmental Data Initiative's (EDI) data portal based on their PASTA identifier. Returns a list of the dataframes and this singular metadata file. To read in the metadata only, see `helpR::read_edi_meta`.
}
\examples{
\dontrun{
# Read in a data product (on bees)
bees <- read_edi(pasta_id = "https://pasta.lternet.edu/package/eml/edi/1210/1", data_type = "csv")

# Check the names
names(bees)
}

}
