% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edi_meta.R
\name{read_edi_meta}
\alias{read_edi_meta}
\title{Read Data Package Metadata from Environmental Data Initiative (EDI) by PASTA Identifier}
\usage{
read_edi_meta(pasta_id = NULL)
}
\arguments{
\item{pasta_id}{(character) PASTA identifier. Can be found in the EDI page for a given data product just above the "Code Generation" buttons for various coding languages.}
}
\value{
(list) containing one element per table in data product and the final element is the XML metadata. List names are the PASTA identifiers for each specific object (note that these differ slightly from the PASTA identifier this function requires)
}
\description{
Reads in only the metadata (as an XML) stored on the Environmental Data Initiative's (EDI) data portal based on the data package's PASTA identifier. Returns only the metadata file. To read in the data and metadata, see `helpR::read_edi`.
}
\examples{
# Read in a data product (on bees)
bees_meta <- read_edi_meta(pasta_id = "https://pasta.lternet.edu/package/eml/edi/1210/1")

# Check the structure
str(bees_meta)

}
