% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{diff_chk}
\alias{date_chk}
\alias{multi_date_chk}
\alias{num_chk}
\alias{multi_num_chk}
\title{Functions renamed in helpR 1.0.0}
\usage{
diff_chk(old = NULL, new = NULL, sort = TRUE, return = FALSE)

date_chk(data = NULL, col = NULL)

multi_date_chk(data = NULL, col_vec = NULL)

num_chk(data = NULL, col = NULL)

multi_num_chk(data = NULL, col_vec = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

helpR 1.0.0 renamed all functions ending in "_chk" to avoid ambiguous function shorthand. These functions all now end in "_check"

\code{diff_chk()} -> \code{diff_check()}

\code{date_chk()} -> \code{date_check()}

\code{multi_date_chk()} -> \code{multi_date_check()}

\code{num_chk()} -> \code{num_check()}

\code{multi_num_chk()} -> \code{multi_num_check()}
}
\keyword{internal}
