% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_export.R
\name{rmd_export}
\alias{rmd_export}
\title{Knit an R Markdown File and Export to Google Drive}
\usage{
rmd_export(
  in_path = getwd(),
  in_name,
  out_path = getwd(),
  out_name,
  out_type = "html",
  drive_link
)
}
\arguments{
\item{in_path}{path to the .Rmd file (defaults to `getwd()`)}

\item{in_name}{name of .Rmd file (with or without ".Rmd" at end)}

\item{out_path}{path to the knit file's destination (defaults to `getwd()`)}

\item{out_name}{desired name for knit file (with or without file suffix)}

\item{out_type}{either "html" or "pdf" depending on what YAML entry you have in the `output: ` field of your .Rmd}

\item{drive_link}{full URL of drive folder to upload the knit document}
}
\description{
This function allows you to knit a specified R Markdown file locally and export it to the Google Drive folder for which you provided a link. NOTE that if you have not used `googledrive::drive_auth()` this will prompt you to authorize a Google account in a new browser tab. If you do not check the box in that screen before continuing you will not be able to use this function until you clear your browser cache and re-authenticate. I recommend invoking `drive_auth()` beforehand to reduce the chances of this error
}
\examples{
# Authorize R to interact with GoogleDrive
# googledrive::drive_auth()
## NOTE: See above warning about possible misstep at this stage

# Use `rmd_export()` to knit and export an .Rmd file
# rmd_export(in_name = "my_markdown.Rmd",
# in_path = file.path("Folder in my WD with the .Rmd named in `in_name`"),
# out_path = file.path("Folder in my WD to save the knit file to"),
# out_name = "desired name for output",
# out_type = "html",
# drive_link = "https://...Google Drive link...")

}
