% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spellcheck_quarto.r
\name{spellcheck_quarto}
\alias{spellcheck_quarto}
\title{Check the Spelling of All Quarto Files in a Given Directory}
\usage{
spellcheck_quarto(path = getwd(), quiet = FALSE)
}
\arguments{
\item{path}{(character) directory in which all Quarto files are to be identified and spell-checked. Recursively identifies Quarto files in folders within the top-level folder. Unrendered folders are ignored. Defaults to current working directory}

\item{quiet}{(logical) whether to message the name of the current file being spell-checked as it is checked. Defaults to \code{FALSE}}
}
\value{
(dataframe) table containing a column for the Quarto file, its path, words flagged as typos by \code{spell_check_files}, and what line(s) of the file that typo was found. One row per typo. If found on multiple lines in a single file, concatenated with comma as separators.
}
\description{
Runs \code{spelling::spell_check_files} on all Quarto (.qmd) files in folders that would be rendered when rendering a Quarto project (i.e., folders that do not begin with an underscore).
}
