% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_data.r
\name{standardize_data}
\alias{standardize_data}
\alias{standardize_data.data.frame}
\title{Standardize data}
\usage{
standardize_data(data, center = TRUE, scale = TRUE, ...)

\method{standardize_data}{data.frame}(data, center = TRUE, scale = TRUE, ...)
}
\arguments{
\item{data}{\code{data.frame} that contains the data for standardization}

\item{center}{Value \code{TRUE} or \code{FALSE} that is passed to function \code{scale}}

\item{scale}{Value \code{TRUE} or \code{FALSE} that is passed to function \code{scale}}
}
\value{
\code{data.frame} or \code{tibble} (when package \code{tibble} is available)
object containing all varibales from the input \code{data.frame} where all numeric
variables are scaled according to \code{base} and \code{scale}.
}
\description{
Standarize (center and scale) all numeric variables in a data.frame.
}
\details{
The function applies the \code{base} function \code{scale} to all numeric variables
in the input \code{data.frame}. When the input options \code{base} and \code{scale}
do not receive logical values as input, the option defaults to \code{TRUE}.
}
\examples{
data <- create_data()
standardize_data(data = data, center = TRUE, scale = TRUE)
standardize_data(data = data, center = TRUE, scale = FALSE)
}
\seealso{
\code{\link[base]{scale}}
}
