% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportR-package.R
\docType{package}
\name{supportR-package}
\alias{supportR}
\alias{supportR-package}
\title{supportR: Support Functions for Wrangling and Visualization}
\description{
Suite of helper functions for data wrangling and visualization. The only theme for these functions is that they tend towards simple, short, and narrowly-scoped. These functions are built for tasks that often recur but are not large enough in scope to warrant an ecosystem of interdependent functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/njlyon0/supportR}
  \item \url{https://njlyon0.github.io/supportR/}
  \item Report bugs at \url{https://github.com/njlyon0/supportR/issues}
}

}
\author{
\strong{Maintainer}: Nicholas J Lyon \email{nickjlyon@gmail.com} (\href{https://orcid.org/0000-0003-3905-1078}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicholas J Lyon \email{nickjlyon@gmail.com} (\href{https://orcid.org/0000-0003-3905-1078}{ORCID}) [copyright holder]
}

}
\keyword{internal}
