% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_rnorm.r
\name{trunc_rnorm}
\alias{trunc_rnorm}
\title{Truncated normal distribution}
\usage{
trunc_rnorm(n, mean = 0, sd = 1, lwr = -Inf, upr = Inf, nnorm = n)
}
\arguments{
\item{n}{number of observations in distribution}

\item{mean}{mean of distribution}

\item{sd}{standard deviation of distribution}

\item{lwr}{lower boundary of distribution}

\item{upr}{upper boundary of distribution}

\item{nnorm}{size of normal distribution from where
\code{n} values are sampled}
}
\value{
vector of values from generated distribution
}
\description{
Generate a normal distribtion that is truncated at the lower and/or upper end.
}
\details{
The function generates a normal distribution for the given parameters
(\code{mean}, \code{sd}) and samples \code{n} values that fall within the
\code{lwr} and \code{upr} bound. If the distribution does not include
\code{n} values within these boundaries the function generates a slightly
larger normal distribution from where it samples \code{n} values.
}
\examples{
trunc_rnorm(50, mean = 1, sd = 0.5, lwr = -0.1, upr = 2)
}
