% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderstatistics.R
\name{WeibullOrder}
\alias{WeibullOrder}
\title{Moments of order statistics for three-parameter Weibull}
\usage{
WeibullOrder(q = 1, r = 1, shape = 1, scale = 1, location = 0, pow = 1)
}
\arguments{
\item{q}{The order of statistics (q-th out of r)}

\item{r}{The number of replicates to extract the q-th statistics from (q-th out of r)}

\item{shape}{The shape of three-parameter Weibull distribution (1 is for exponential)}

\item{scale}{The scale of three-parameter Weibull distribution}

\item{location}{The location of three-parameter Weibull distribution (0 default)}

\item{pow}{The order of the moment to obtain statistics for (default pow=1 stands for expectation)}
}
\value{
The pow-th moment of q-th order statistics of r iid. copies of the
   r.v. distribued as the three-parameter Weibull with distribution function
   \deqn{F(x)=1-exp\left(-\left( \frac{x-location}{scale}\right)^{shape} \right).}
}
\description{
WeibullOrder is the moments of order statistics from three-parameter Weibull distribution.
}
\examples{
WeibullOrder() # should be 1 as the expectation of a standard Exponential
WeibullOrder(1,10) # should be 0.1 as the minimum of 10 standard Exponentially distributed r.v.'s
}
\references{
C.-D. Lai, Generalized Weibull Distributions. Springer: 2014. doi: 10.1007/978-3-642-39106-4.
   J. Lieblein, “On Moments of Order Statistics from the Weibull Distribution,” Ann. Math. Statist., vol. 26, no. 2, pp. 330–333, Jun. 1955, doi: 10.1214/aoms/1177728551.
}
\seealso{
[supportR::ParetoOrder()] for moments from Pareto distribution.
}
