% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR-package.R
\docType{package}
\name{helpR-package}
\alias{helpR}
\alias{helpR-package}
\title{helpR: Miscellaneous Helper Functions for Coding in R}
\description{
Suite of helper functions for data wrangling and visualization. The only theme for these functions is that they tend towards simple, short, and narrowly-defined. These functions are built for tasks that often recur but are not large enough in scope to warrant an ecosystem of inter-dependent functions.
}
\author{
\strong{Maintainer}: Nicholas J Lyon \email{njlyon@alumni.iastate.edu} (\href{https://orcid.org/0000-0003-3905-1078}{ORCID})

}
\keyword{internal}
