% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_to_prob.r
\name{logit_to_prob}
\alias{logit_to_prob}
\title{Logit to probability}
\usage{
logit_to_prob(logit)
}
\arguments{
\item{logit}{\code{numeric} vector of logit values}
}
\value{
Probability value of the logit value served as input.
}
\description{
The function transforms a logit value to a probability value.
}
\examples{
data <- create_data()
mod <- glm(female_ceo ~ advertising, data = data, family = "binomial")
coefs <- coef(mod)
logit_to_prob(logit = coefs[["advertising"]])
}
\references{
\itemize{
  \item \url{https://en.wikipedia.org/wiki/Logit}
  \item \url{https://sebastiansauer.github.io/Rcode/logit2prob.R}
}
}
