% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lyon.R
\name{theme_lyon}
\alias{theme_lyon}
\title{Complete \code{ggplot2} Theme for Non-Data Aesthetics}
\usage{
theme_lyon(title_size = 16, text_size = 13)
}
\arguments{
\item{title_size}{(numeric) size of font in axis titles}

\item{text_size}{(numeric) size of font in tick labels}
}
\value{
(ggplot theme) list of ggplot2 theme elements
}
\description{
Custom alternative to the \code{ggtheme} options built into \code{ggplot2}. Removes gray boxes and grid lines from plot background. Increases font size of tick marks and axis labels. Removes gray box from legend background and legend key. Removes legend title.
}
